spanner/spansql: handle quoted identifiers in parsing and formatting
This accepts identifiers enclosed by backtics, which are required for
keywords and other special names, and formats identifiers that are
keywords when generating SQL.

This does not affect case insensitivity, nor tighten parsing (e.g. the
query "SELECT Foo `FROM` Tbl" shouldn't be accepted), but covers an
important use case, namely tables and columns that have names that are
reserved keywords.

Fixes #1710.

Change-Id: Ibe1468f594f1abbf76739016073ba873b9d2599c
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/51415
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Knut Olav Løite <koloite@gmail.com>